\name{createalldata}
\alias{createalldata}
\title{Loops through data creation}
\usage{
  createalldata(alldata, ngrps, cc_dir, cc_rfile)
}
\arguments{
  \item{alldata}{list generated by getdata() of all lines
  expected to contain data}

  \item{ngrps}{number of unique solution sets to produce}

  \item{cc_dir}{location of output directory to be written
  to during replication process}

  \item{cc_rfile}{location of output file to be written to
  during replication process}
}
\description{
  Main data creation function that generates folders for
  data storage and then loops through each line given by
  getdata() and generates a new data frame for each group.
}

