\name{createcamconfolder}
\alias{createcamconfolder}
\title{Creates folder for camcon files}
\usage{
  createcamconfolder(rfile)
}
\arguments{
  \item{rfile}{name of the R file being replicated through
  camcon}
}
\description{
  Creates folder of format /camcon_filename/ for storing
  data files, .csv file with solutions, and .png files
}

