\name{createdata}
\alias{createdata}
\title{Finds the name and source of the data file}
\usage{
  createdata(data, params, ngrps, cc_dir, cc_rfile)
}
\arguments{
  \item{data}{a single line generated by getdata() that
  contains the data set and specified parameters}

  \item{params}{the parameters identified by
  detectdatatype()}

  \item{ngrps}{number of unique solution sets to produce}

  \item{cc_dir}{location of output directory to be written
  to during replication process}

  \item{cc_rfile}{location of output file to be written to
  during replication process}
}
\description{
  Uses a line given by getdata() and detects the data frame
  name and the source of the data, and generates a new data
  frame for each group depending on the data type.
}

