\name{createparam}
\alias{createparam}
\title{Writes to camcon script a new list of parameters of length equal to the number of groups}
\usage{
  createparam(param, ngrps, cc_rfile)
}
\arguments{
  \item{param}{a single line generated by getparams() that
  contains the parameter}

  \item{ngrps}{number of unique solution sets to produce}

  \item{cc_rfile}{location of output file to be written to
  during replication process}
}
\description{
  Overwrites the original parameter vector with a new
  vector of length equal to the number of groups so that
  individuals groups have original parameter values
}

