\name{createtsdata}
\alias{createtsdata}
\title{Writes to camcon script for time-series data}
\usage{
  createtsdata(data, sd, lhs, rhs, ngrps, cc_dir, cc_rfile)
}
\arguments{
  \item{data}{a single line generated by getdata() that
  contains the data set and specified parameters}

  \item{sd}{the standard deviation for the random noise}

  \item{lhs}{the name of the original data frame}

  \item{rhs}{the input for the original data frame}

  \item{ngrps}{number of unique solution sets to produce}

  \item{cc_dir}{location of output directory to be written
  to during replication process}

  \item{cc_rfile}{location of output file to be written to
  during replication process}
}
\description{
  Generates a new data frame for every individual by adding
  random noise
}

