\name{detectdatatype}
\alias{detectdatatype}
\title{Check whether cross-sectional, time series, or panel data}
\usage{
  detectdatatype(data)
}
\arguments{
  \item{data}{a single line generated by getdata() that
  contains the data set and specified parameters}
}
\description{
  Detects the type of data, and ensures that valid
  associated parameters are supplied
}

