\name{getstrlocs}
\alias{getstrlocs}
\title{Matches location of sections in script}
\usage{
  getstrlocs(script, strpattern)
}
\arguments{
  \item{script}{line by line list produced from R file for
  replication}

  \item{strpattern}{pattern associated with data,
  parameter, and script sections in rfile}
}
\description{
  Creates a vector of matched string locations.
}

