\name{valid_var_names}
\alias{valid_var_names}
\title{Determines whether string is a valid variable name}
\usage{
  valid_var_names(qname)
}
\arguments{
  \item{qname}{string to convert into a valid question
  name}
}
\description{
  Custom function to determine whether the new string
  generated from the question string is a valid variable
  name
}

