\name{camcon_init}
\alias{camcon_init}
\title{Creates folder for camcon files and blank script}
\usage{
  camcon_init(rfile)
}
\arguments{
  \item{rfile}{name of the R file being replicated through
  camcon}
}
\description{
  Creates folder /camcon_filename/ for storing data files,
  question sheets, .csv file with solutions, and .png
  files, and blank script for camcon replication.
}

